+++ Sources/ClingsCLI/Clings.swift
@@ -32,7 +32,7 @@ struct Clings: AsyncParsableCommand {
         For more information on a specific command, run:
           clings <command> --help
         """,
-        version: "0.3.0",
+        version: "0.3.1",
         subcommands: [
             // List views
             TodayCommand.self,